Version 1.6: Allow checking goo.im if pa.co fails

+++ src/com/paranoid/paranoidota/updater/Updater.java
@@ -34,7 +34,7 @@ public abstract class Updater implements URLStringReaderListener, HttpStringRead
         public String getFilename();
         public String getPath();
         public String getSize();
-        public long getVersion();
+        public double getVersion();
         public boolean isDelta();
         public String getDeltaFilename();
         public String getDeltaPath();
